사용하지 않는 vm exe를 위한 프로젝트 삭제, 네이티브 바인딩을 쉽게 처리하기위한 embed 프로젝트 추가(dll) 및 문서내용 추가 및 수정

+++ src/common/ss_compiler.cpp
@@ -350,7 +350,7 @@ void Compiler::visit(ClassDeclStmt* stmt) {
                 will_set_compiler.emit_op(OpCode::OP_RETURN, property->line);
 
                 will_set_proto.chunk = finalize_function_chunk(std::move(will_set_compiler.chunk_));
-                will_set_idx = chunk_.add_function(std::move(will_set_proto));
+                will_set_idx = static_cast<uint16_t>(chunk_.add_function(std::move(will_set_proto)));
             }
 
             // Compile didSet observer if present
@@ -388,7 +388,7 @@ void Compiler::visit(ClassDeclStmt* stmt) {
                 did_set_compiler.emit_op(OpCode::OP_RETURN, property->line);
 
                 did_set_proto.chunk = finalize_function_chunk(std::move(did_set_compiler.chunk_));
-                did_set_idx = chunk_.add_function(std::move(did_set_proto));
+                did_set_idx = static_cast<uint16_t>(chunk_.add_function(std::move(did_set_proto)));
             }
 
             // Emit property definition with observers
@@ -686,7 +686,7 @@ if (scope_depth_ > 0) {
                 will_set_compiler.emit_op(OpCode::OP_RETURN, property->line);
 
                 will_set_proto.chunk = finalize_function_chunk(std::move(will_set_compiler.chunk_));
-                will_set_idx = chunk_.add_function(std::move(will_set_proto));
+                will_set_idx = static_cast<uint16_t>(chunk_.add_function(std::move(will_set_proto)));
             }
 
             // Compile didSet observer if present
@@ -724,7 +724,7 @@ if (scope_depth_ > 0) {
                 did_set_compiler.emit_op(OpCode::OP_RETURN, property->line);
 
                 did_set_proto.chunk = finalize_function_chunk(std::move(did_set_compiler.chunk_));
-                did_set_idx = chunk_.add_function(std::move(did_set_proto));
+                did_set_idx = static_cast<uint16_t>(chunk_.add_function(std::move(did_set_proto)));
             }
 
             // Emit property definition with observers
